Add RSS button to footer, if it is defined in the site's path.rss
RSS is enabled by default on the website using jekyll-feed, but the
majojrity of users would not know how to find it without a link. This
commit provides a "button" at the bottom of the page, like the email,
GitHub, etc. icon-links which will redirect to the RSS XML file location.